Institution Domaene bereinigt

+++ src/main/java/io/remedymatch/institution/api/AnfrageDTO.java
@@ -26,7 +26,7 @@ public class AnfrageDTO {
 
     private InstitutionStandortDTO standortVon;
 
-    private double entfernung;
+    private BigDecimal entfernung;
 
     private String prozessInstanzId;
